+++ src/app/hazo_auth/profile_stamp_test/page.tsx
@@ -1,0 +1,11 @@
+// file_description: Test page for ProfileStamp component
+import { AuthPageShell } from "../../../components/layouts/shared/components/auth_page_shell";
+import { ProfileStampTestLayout } from "../../../components/layouts/profile_stamp_test";
+
+export default function ProfileStampTestPage() {
+  return (
+    <AuthPageShell>
+      <ProfileStampTestLayout />
+    </AuthPageShell>
+  );
+}